requirements.txt
flask==2.3.0
openai==2.15.0
gradio_client==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.0
